Commence transition to /src layout with edubot.py
This commit begins the process of transitioning to the /src layout which
is beneficial for installationa and testing purposes as the installed
code is always the one being tested. This first commit moves the
edubot.py module into /src/edubot/bot.py. Note that the renaming of this
Python module is done so that it does not conflict with the name of the
package.

Summary of Changes:
- Moving/renaming edubot.py to src/edubot/bot.py
- Adding AGPL-3.0 license header to bot.py
- Running black formatter on file
- Making docstrings conform to Google docstring guidelines
